requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.6.1
Django==2.2.1
django-cors-headers==3.0.1
django-rest-knox==4.0.1
djangorestframework==3.9.4
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
